backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
pytest
httpx
